requirements.txt
colorama==0.4.6
azure-functions==1.14.0
